Support of partial reconfiguration and other features.
fixed issue:
msg folder missing in ros_sort_demo #6

+++ tools/_pypack/reconos/scripts/msg/export.py
@@ -33,52 +33,56 @@ def export_msg_cmd(args):
 	
 def export_msg(args, msgdir, link):
 	prj = args.prj
-	msgdir = msgdir if msgdir is not None else prj.basedir + ".msg"
+	if not os.path.exists(prj.dir + "/msg/"):
+		print("No msg folder, nothing to do.")
+	else:
+		
+		msgdir = msgdir if msgdir is not None else prj.basedir + ".msg"
 
-	if shutil2.exists(msgdir):
-		shutil.rmtree(msgdir)
-	shutil2.mkdir(msgdir)
+		if shutil2.exists(msgdir):
+			shutil.rmtree(msgdir)
+		shutil2.mkdir(msgdir)
 
-	log.info("Export software to project directory '" + prj.dir + "'")
+		log.info("Export software to project directory '" + prj.dir + "'")
 
-	msg_packages = [f for f in listdir(prj.dir + "/msg/") if isdir(join(prj.dir + "/msg/", f))]
+		msg_packages = [f for f in listdir(prj.dir + "/msg/") if isdir(join(prj.dir + "/msg/", f))]
 
-	print(msg_packages)
+		print(msg_packages)
 
-	for msgs_packs in msg_packages:
-		dictionary = {}
-		dictionary["pkgname"] = msgs_packs
-		srv_files = {}
-		if shutil2.exists(prj.dir + "/msg/"+msgs_packs+"/srv"):
-			srv_files = [f for f in listdir(prj.dir + "/msg/"+msgs_packs+"/srv") if isfile(join(prj.dir + "/msg/"+msgs_packs+"/srv", f))]
-			print(srv_files)
-		msg_files = {}
-		if shutil2.exists(prj.dir + "/msg/"+msgs_packs+"/msg/"):
-			msg_files = [f for f in listdir(prj.dir + "/msg/"+msgs_packs+"/msg/") if isfile(join(prj.dir + "/msg/"+msgs_packs+"/msg/", f))]
-			print(msg_files)
-		action_files = {}
-		if shutil2.exists(prj.dir + "/msg/"+msgs_packs+"/action/"):
-			action_files = [f for f in listdir(prj.dir + "/msg/"+msgs_packs+"/action/") if isfile(join(prj.dir + "/msg/"+msgs_packs+"/action/", f))]
-			print(action_files)
+		for msgs_packs in msg_packages:
+			dictionary = {}
+			dictionary["pkgname"] = msgs_packs
+			srv_files = {}
+			if shutil2.exists(prj.dir + "/msg/"+msgs_packs+"/srv"):
+				srv_files = [f for f in listdir(prj.dir + "/msg/"+msgs_packs+"/srv") if isfile(join(prj.dir + "/msg/"+msgs_packs+"/srv", f))]
+				print(srv_files)
+			msg_files = {}
+			if shutil2.exists(prj.dir + "/msg/"+msgs_packs+"/msg/"):
+				msg_files = [f for f in listdir(prj.dir + "/msg/"+msgs_packs+"/msg/") if isfile(join(prj.dir + "/msg/"+msgs_packs+"/msg/", f))]
+				print(msg_files)
+			action_files = {}
+			if shutil2.exists(prj.dir + "/msg/"+msgs_packs+"/action/"):
+				action_files = [f for f in listdir(prj.dir + "/msg/"+msgs_packs+"/action/") if isfile(join(prj.dir + "/msg/"+msgs_packs+"/action/", f))]
+				print(action_files)
 
-		dictionary["interface_files"] = ""
-		for f in srv_files:
-			dictionary["interface_files"] += '"srv/'+f+'" '
-	
-		for f in msg_files:
-			dictionary["interface_files"] += '"msg/'+f+'" '
+			dictionary["interface_files"] = ""
+			for f in srv_files:
+				dictionary["interface_files"] += '"srv/'+f+'" '
+		
+			for f in msg_files:
+				dictionary["interface_files"] += '"msg/'+f+'" '
 
-		for f in action_files:
-			dictionary["interface_files"] += '"action/'+f+'" '
+			for f in action_files:
+				dictionary["interface_files"] += '"action/'+f+'" '
 
-		print(dictionary["interface_files"])	
-		prj.apply_template("ros_msg", dictionary, msgdir + "/"+ msgs_packs, link)
-		if shutil2.exists(prj.dir + "/msg/" + msgs_packs + "/srv/"):
-			shutil.copytree( prj.dir + "/msg/" + msgs_packs + "/srv/", msgdir + "/" + msgs_packs + "/srv/")
-		if shutil2.exists(prj.dir + "/msg/" + msgs_packs + "/msg/"):	
-			shutil.copytree( prj.dir + "/msg/" + msgs_packs + "/msg/", msgdir + "/" + msgs_packs + "/msg/")
-		if shutil2.exists(prj.dir + "/msg/" + msgs_packs + "/action/"):	
-			shutil.copytree( prj.dir + "/msg/" + msgs_packs + "/action/", msgdir + "/" + msgs_packs + "/action/")
+			print(dictionary["interface_files"])	
+			prj.apply_template("ros_msg", dictionary, msgdir + "/"+ msgs_packs, link)
+			if shutil2.exists(prj.dir + "/msg/" + msgs_packs + "/srv/"):
+				shutil.copytree( prj.dir + "/msg/" + msgs_packs + "/srv/", msgdir + "/" + msgs_packs + "/srv/")
+			if shutil2.exists(prj.dir + "/msg/" + msgs_packs + "/msg/"):	
+				shutil.copytree( prj.dir + "/msg/" + msgs_packs + "/msg/", msgdir + "/" + msgs_packs + "/msg/")
+			if shutil2.exists(prj.dir + "/msg/" + msgs_packs + "/action/"):	
+				shutil.copytree( prj.dir + "/msg/" + msgs_packs + "/action/", msgdir + "/" + msgs_packs + "/action/")
 
 	# #bashCommand = "source /opt/ros/foxy/setup.bash; cd "+swdir+ "/msg; ros2 pkg create my_reconros_services --dependencies builtin_interfaces"
 	# #print(bashCommand)
